+++ src/server/game/Handlers/AuctionHouseHandler.cpp
@@ -69,7 +69,7 @@ void WorldSession::SendAuctionHello(ObjectGuid guid, Unit const* unit)
         return;
 
     WorldPacket data(MSG_AUCTION_HELLO, 12);
-    data << uint64(guid);
+    data << guid;
     data << uint32(ahEntry->ID);
     data << uint8(1);                                       // 3.3.3: 1 - AH enabled, 0 - AH disabled
     SendPacket(&data);
@@ -110,7 +110,7 @@ void WorldSession::SendAuctionBidderNotification(uint32 location, uint32 auction
     WorldPacket data(SMSG_AUCTION_BIDDER_NOTIFICATION, (8*4));
     data << uint32(location);
     data << uint32(auctionId);
-    data << uint64(bidder);
+    data << bidder;
     data << uint32(bidSum);
     data << uint32(diff);
     data << uint32(itemEntry);
